% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalz.R
\name{C_db}
\alias{C_db}
\title{Finalisation des calculs et Exportation des Outputs}
\usage{
C_db(yearz)
}
\arguments{
\item{yearz}{}
}
\value{
les datasets contenant les indices et les outputs dans le dossier eiti
}
\description{
Remplissage avec le prix de la période la plus proche pour Laspeyres, Aucune imputation pour Paasche,
}
\examples{

}
