% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charge.R
\name{charge_dp}
\alias{charge_dp}
\title{Chargement de la base et des dictionnaires}
\usage{
charge_dp(dbc, dom, dp, syear, eyear)
}
\arguments{
\item{dbc, dp, syear, eyear}{}
}
\value{
Dataframes contenant une sélection du dataset principal et des dictionnaires/relations
}
\description{
Charge de dataset principal, les dictionnaires et les relations
}
\examples{

}
