% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outltukmmiriq.R
\name{OutlTuk3}
\alias{OutlTuk3}
\title{Fonctions usuelles de calcul des Indicateurs MMI RIQ et détection d'outliers}
\usage{
OutlTuk3(dt, var, n)
}
\arguments{
\item{input_df, customsoffice, modeoftransport, cpc, partner, product, value, netweight, quantity}{}
}
\value{
Le dataset BJC
}
\description{
OutlTuk et MMIRIQ
}
\examples{

}
