% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_BJC.R
\name{prep_db}
\alias{prep_db}
\title{Renomme champs et Calcul de la Valeur Unitaire}
\usage{
prep_db(input_df, customsoffice, modeoftransport, cpc, partner, product, value,
  netweight, quantity)
}
\arguments{
\item{input_df, customsoffice, modeoftransport, cpc, partner, product, value, netweight, quantity}{}
}
\value{
Le dataset BJC
}
\description{
Renomme les champs Bureau de douane, mode de transport, régime, partenaire, produit HS, valeur, poids net et quantité et crée le champ valeur unitaire. Il est nécessaire que les dictionnaires et relations Flux et Type de Commerce soient au préalable chargés.
}
\examples{

}
